+++ Source/MidiHandler.cpp
@@ -7,12 +7,13 @@
 #include "ApplicationSettings.h"
 #include "errorhandlers.h"
 #include "VersionToken.h"
+#include "ProgramHeader.h"
+#include "PatchRegistry.h"
 #ifdef USE_CODEC
 #include "Codec.h"
 #endif
 #include "Owl.h"
-#include "FlashStorage.h"
-#include "PatchRegistry.h"
+#include "Storage.h"
 #ifndef USE_BOOTLOADER_MODE
 #include "BootloaderStorage.h"
 #endif
@@ -33,22 +34,11 @@ MidiHandler::MidiHandler() : channel(MIDI_OMNI_CHANNEL) {
   // memset(midi_values, 0, NOF_PARAMETERS*sizeof(uint16_t));
 }
 
-void MidiHandler::handlePitchBend(uint8_t status, uint16_t value){
-}
+void MidiHandler::handlePitchBend(uint8_t status, uint16_t value){}
 
-void MidiHandler::handleNoteOn(uint8_t status, uint8_t note, uint8_t velocity){
-  if(channel != MIDI_OMNI_CHANNEL && channel != getChannel(status))
-    return;
-  if(getProgramVector()->buttonChangedCallback != NULL)
-    getProgramVector()->buttonChangedCallback(MIDI_NOTE_BUTTON+note, velocity<<5, getSampleCounter());
-}
+void MidiHandler::handleNoteOn(uint8_t status, uint8_t note, uint8_t velocity){}
 
-void MidiHandler::handleNoteOff(uint8_t status, uint8_t note, uint8_t velocity){
-  if(channel != MIDI_OMNI_CHANNEL && channel != getChannel(status))
-    return;
-  if(getProgramVector()->buttonChangedCallback != NULL)
-    getProgramVector()->buttonChangedCallback(MIDI_NOTE_BUTTON+note, 0, getSampleCounter());
-}
+void MidiHandler::handleNoteOff(uint8_t status, uint8_t note, uint8_t velocity){}
 
 void MidiHandler::handleProgramChange(uint8_t status, uint8_t pid){
   if(channel != MIDI_OMNI_CHANNEL && channel != getChannel(status))
@@ -94,6 +84,23 @@ void MidiHandler::handleControlChange(uint8_t status, uint8_t cc, uint8_t value)
   case MIDI_CC_EFFECT_CTRL_2:
     setParameterValue(PARAMETER_H, value<<5);
     break;
+#ifdef USE_ADC
+  case PATCH_CONTROL:
+    /* Remote control: 0=local, 127=MIDI */
+    if(value){
+      extern ADC_HandleTypeDef ADC_PERIPH;
+      extern uint16_t adc_values[NOF_ADC_VALUES];
+      HAL_StatusTypeDef ret = HAL_ADC_Start_DMA(&ADC_PERIPH, (uint32_t*)adc_values, NOF_ADC_VALUES);
+      if(ret != HAL_OK)
+	error(CONFIG_ERROR, "ADC Start failed");
+    }else{
+      extern ADC_HandleTypeDef ADC_PERIPH;
+      HAL_StatusTypeDef ret = HAL_ADC_Stop_DMA(&ADC_PERIPH);
+      if(ret != HAL_OK)
+	error(CONFIG_ERROR, "ADC Stop failed");
+    }
+    break;
+#endif
   case PATCH_BUTTON:
     setButtonValue(PUSHBUTTON, value < 64 ? 0 : 255);
     break;
@@ -111,9 +118,6 @@ void MidiHandler::handleControlChange(uint8_t status, uint8_t cc, uint8_t value)
       break;
     case SYSEX_PRESET_NAME_COMMAND:
       midi_tx.sendPatchNames();
-      break;
-    case SYSEX_PARAMETER_NAME_COMMAND:
-      midi_tx.sendPatchParameterNames();
       break;
     case SYSEX_CONFIGURATION_COMMAND:
       midi_tx.sendSettings();
@@ -173,61 +177,83 @@ void MidiHandler::updateCodecSettings(){
   program.resetProgram(true);
 }
 
+#define HEXCODE(x) ((uint16_t)(((x)[1]<<8)|(x)[0]))
+
 void MidiHandler::handleConfigurationCommand(uint8_t* data, uint16_t size){
   if(size < 3) // size may be 3 or more depending on number of digits in value
     return;
-  char* p = (char*)data;
-  int32_t value = strtol(p+2, NULL, 16);
-  if(strncmp(SYSEX_CONFIGURATION_AUDIO_RATE, p, 2) == 0){
+  uint16_t hex = HEXCODE(data);
+  int32_t value = strtol((char*)data+2, NULL, 16);
+  switch(hex){
+  case HEXCODE(SYSEX_CONFIGURATION_AUDIO_RATE):
     settings.audio_samplingrate = value;
-  }else if(strncmp(SYSEX_CONFIGURATION_AUDIO_BLOCKSIZE, p, 2) == 0){
+    break;
+  case HEXCODE(SYSEX_CONFIGURATION_AUDIO_BLOCKSIZE):
     settings.audio_blocksize = value;
     updateCodecSettings();
-  }else if(strncmp(SYSEX_CONFIGURATION_AUDIO_BITDEPTH, p, 2) == 0){
+    break;
+  case HEXCODE(SYSEX_CONFIGURATION_AUDIO_BITDEPTH):
     settings.audio_bitdepth = value;
-  }else if(strncmp(SYSEX_CONFIGURATION_AUDIO_DATAFORMAT, p, 2) == 0){
+    break;
+  case HEXCODE(SYSEX_CONFIGURATION_AUDIO_DATAFORMAT):
     settings.audio_dataformat = value;
+    break;
 #ifdef USE_CODEC
-  }else if(strncmp(SYSEX_CONFIGURATION_CODEC_SWAP, p, 2) == 0){
+  case HEXCODE(SYSEX_CONFIGURATION_CODEC_SWAP):
     settings.audio_codec_swaplr = value;
-  }else if(strncmp(SYSEX_CONFIGURATION_CODEC_BYPASS, p, 2) == 0){
+    break;
+  case HEXCODE(SYSEX_CONFIGURATION_CODEC_BYPASS):
     settings.audio_codec_bypass = value;
     codec.bypass(value);
-  }else if(strncmp(SYSEX_CONFIGURATION_CODEC_INPUT_GAIN, p, 2) == 0){
+    break;
+  case HEXCODE(SYSEX_CONFIGURATION_CODEC_INPUT_GAIN):
     settings.audio_input_gain = value;  
     codec.setInputGain(settings.audio_input_gain);
-  }else if(strncmp(SYSEX_CONFIGURATION_CODEC_OUTPUT_GAIN, p, 2) == 0){
+    break;
+  case HEXCODE(SYSEX_CONFIGURATION_CODEC_OUTPUT_GAIN):
     settings.audio_output_gain = value;  
     codec.setOutputGain(settings.audio_output_gain);
-  }else if(strncmp(SYSEX_CONFIGURATION_CODEC_HIGHPASS, p, 2) == 0){
+    break;
+  case HEXCODE(SYSEX_CONFIGURATION_CODEC_HIGHPASS):
     codec.setHighPass(value);
+    break;
 #endif
-  }else if(strncmp(SYSEX_CONFIGURATION_PC_BUTTON, p, 2) == 0){
+  case HEXCODE(SYSEX_CONFIGURATION_PC_BUTTON):
     settings.program_change_button = value;
-  }else if(strncmp(SYSEX_CONFIGURATION_INPUT_OFFSET, p, 2) == 0){
+    break;
+  case HEXCODE(SYSEX_CONFIGURATION_INPUT_OFFSET):
     settings.input_offset = value;
-  }else if(strncmp(SYSEX_CONFIGURATION_INPUT_SCALAR, p, 2) == 0){
+    break;
+  case HEXCODE(SYSEX_CONFIGURATION_INPUT_SCALAR):
     settings.input_scalar = value;
-  }else if(strncmp(SYSEX_CONFIGURATION_OUTPUT_OFFSET, p, 2) == 0){
+    break;
+  case HEXCODE(SYSEX_CONFIGURATION_OUTPUT_OFFSET):
     settings.output_offset = value;
-  }else if(strncmp(SYSEX_CONFIGURATION_OUTPUT_SCALAR, p, 2) == 0){
+    break;
+  case HEXCODE(SYSEX_CONFIGURATION_OUTPUT_SCALAR):
     settings.output_scalar = value;
-  }else if(strncmp(SYSEX_CONFIGURATION_MIDI_INPUT_CHANNEL, p, 2) == 0){
+    break;
+  case HEXCODE(SYSEX_CONFIGURATION_MIDI_INPUT_CHANNEL):
     midiSetInputChannel(max(-1, min(15, value)));
-  }else if(strncmp(SYSEX_CONFIGURATION_MIDI_OUTPUT_CHANNEL, p, 2) == 0){
+    break;
+  case HEXCODE(SYSEX_CONFIGURATION_MIDI_OUTPUT_CHANNEL):
     midiSetOutputChannel(max(-1, min(15, value)));
+    break;
 #ifndef USE_BOOTLOADER_MODE
-  }else if(strncmp(SYSEX_CONFIGURATION_BOOTLOADER_LOCK, p, 2) == 0){
+  case HEXCODE(SYSEX_CONFIGURATION_BOOTLOADER_LOCK):
     if (value)
       bootloader.lock();
     else
       bootloader.unlock();
+    break;
 #endif
 #ifdef USE_DIGITALBUS
-  }else if(strncmp(SYSEX_CONFIGURATION_BUS_ENABLE, p, 2) == 0){
+  case HEXCODE(SYSEX_CONFIGURATION_BUS_ENABLE):
     settings.bus_enabled = value;
-  }else if(strncmp(SYSEX_CONFIGURATION_BUS_FORWARD_MIDI, p, 2) == 0){
+    break;
+  case HEXCODE(SYSEX_CONFIGURATION_BUS_FORWARD_MIDI):
     settings.bus_forward_midi = value;
+    break;
 #endif
   }
 }
@@ -245,7 +271,7 @@ void MidiHandler::handleFirmwareUploadCommand(uint8_t* data, uint16_t size){
   int32_t ret = loader.handleFirmwareUpload(data, size);
   if(ret == 0){
     owl.setOperationMode(LOAD_MODE);
-    setParameterValue(LOAD_INDICATOR_PARAMETER, loader.index*4095/loader.size);
+    setProgress((uint64_t)loader.getLoadedSize()*4095/loader.getDataSize(), "Loading");
   }
 }
 
@@ -255,8 +281,7 @@ void MidiHandler::handleFirmwareRunCommand(uint8_t* data, uint16_t size){
 
 void MidiHandler::runProgram(){
   if(loader.isReady()){
-    // program.exitProgram(true); // exit progress bar
-    program.loadDynamicProgram(loader.getData(), loader.getSize());
+    program.loadDynamicProgram(loader.getData(), loader.getDataSize());
     loader.clear();
     // program.startProgram(true);
     program.resetProgram(true);
@@ -293,8 +318,7 @@ void MidiHandler::handleFirmwareFlashCommand(uint8_t* data, uint16_t size){
       }
       else {
         //program.eraseFromFlash(-2);
-        program.saveToFlash(-2, loader.getData(), loader.getSize());
-        loader.clear();
+        program.saveToFlash(-2, loader.getData(), loader.getDataSize());
         program.resetProgram(true);
       }
     }else{
@@ -303,20 +327,44 @@ void MidiHandler::handleFirmwareFlashCommand(uint8_t* data, uint16_t size){
   }else{
     error(PROGRAM_ERROR, "Invalid FLASH command");
   }
+  loader.clear();
+}
+
+void MidiHandler::handleFirmwareSendCommand(uint8_t* data, uint16_t size){
+  uint32_t slot = loader.decodeInt(data);
+  Resource* resource = NULL;
+  if(slot-1 < MAX_NUMBER_OF_PATCHES-1)
+    resource = registry.getPatch(slot-1);
+  else if(slot-MAX_NUMBER_OF_PATCHES < MAX_NUMBER_OF_RESOURCES)
+    resource = registry.getResource(slot-MAX_NUMBER_OF_PATCHES);
+  if(resource != NULL)
+    midi_tx.sendResource(resource);
+    // program.sendResource(resource);
+  else
+    error(PROGRAM_ERROR, "Invalid SEND command");
 }
 
 void MidiHandler::handleFirmwareStoreCommand(uint8_t* data, uint16_t size){
   if(loader.isReady() && size == 5){
     uint32_t slot = loader.decodeInt(data);
-    if(slot > 0 && slot <= MAX_NUMBER_OF_PATCHES+MAX_NUMBER_OF_RESOURCES){
-      program.saveToFlash(slot, loader.getData(), loader.getSize());
-      loader.clear();
+    if(slot > 0 && slot <= MAX_NUMBER_OF_PATCHES){
+      data = (uint8_t*)loader.getResourceHeader();
+      size_t datasize = loader.getDataSize();
+      ProgramHeader* header = (ProgramHeader*)loader.getData();
+      if(header->magic == 0XDADAC0DE){	
+	storage.writeResourceHeader(data, header->programName, datasize,
+				    FLASH_DEFAULT_FLAGS|RESOURCE_USER_PATCH|slot);
+	program.saveToFlash(slot, data, loader.getTotalSize());
+      }else{
+	error(PROGRAM_ERROR, "Invalid patch magic");
+      }
     }else{
       error(PROGRAM_ERROR, "Invalid STORE slot");
     }
   }else{
     error(PROGRAM_ERROR, "Invalid STORE command");
   }
+  loader.clear();
 }
 
 void MidiHandler::handleFirmwareSaveCommand(uint8_t* data, uint16_t size){
@@ -324,32 +372,17 @@ void MidiHandler::handleFirmwareSaveCommand(uint8_t* data, uint16_t size){
     const char* name = (const char*)data;
     size_t len = strnlen(name, 20);
     if(len > 0 && len < 20){
-      // todo: create ResourceHeader in FirmwareLoader::beginFirmwareUpload()
-      // stop patch or check if running
-      // flash in background task
-      uint32_t slot;
-      ResourceHeader* res = registry.getResource(name);
-      if(res == NULL)
-	slot = registry.getNumberOfResources()+MAX_NUMBER_OF_PATCHES+1;
-      else
-	slot = registry.getSlot(res);
-      data = loader.getData();
-      size_t datasize = loader.getSize();
-      memmove(data+sizeof(ResourceHeader), data, datasize); // make space for resource header
-      memset(data, 0, sizeof(ResourceHeader)); // zero fill header
-      res = (ResourceHeader*)data;
-      res->magic = 0xDADADEED;
-      res->size = datasize;
-      strcpy(res->name, name);
-      datasize += sizeof(ResourceHeader);
-      program.saveToFlash(slot, data, datasize);
-      loader.clear();
+      data = (uint8_t*)loader.getResourceHeader();
+      size_t datasize = loader.getDataSize();
+      storage.writeResourceHeader(data, name, datasize, FLASH_DEFAULT_FLAGS);
+      program.saveToFlash(0, data, loader.getTotalSize());
     }else{
       error(PROGRAM_ERROR, "Invalid SAVE name");
     }
   }else{
     error(PROGRAM_ERROR, "Invalid SAVE command");
   }
+  loader.clear();
 }
 
 void MidiHandler::handleSysEx(uint8_t* data, uint16_t size){
@@ -376,6 +409,9 @@ void MidiHandler::handleSysEx(uint8_t* data, uint16_t size){
     break;
   case SYSEX_FIRMWARE_STORE:
     handleFirmwareStoreCommand(data+4, size-5);
+    break;
+  case SYSEX_FIRMWARE_SEND:
+    handleFirmwareSendCommand(data+4, size-5);
     break;
   case SYSEX_FIRMWARE_FLASH:
     handleFirmwareFlashCommand(data+4, size-5);