+++ Source/PatchRegistry.h
@@ -3,8 +3,9 @@
 
 #include "device.h"
 #include "PatchDefinition.hpp"
-#include "StorageBlock.h"
-#include "ResourceHeader.h"
+#include "Resource.h"
+// #include "StorageBlock.h"
+// #include "ResourceHeader.h"
 
 class PatchRegistry;
 extern PatchRegistry registry;
@@ -16,30 +17,28 @@ public:
   /* const char* getName(unsigned int index); */
   const char* getPatchName(unsigned int index);
   const char* getResourceName(unsigned int index);
-  PatchDefinition* getPatchDefinition(unsigned int index);
+  PatchDefinition* getPatchDefinition(){
+    return &patchDefinition;
+  }
   unsigned int getNumberOfPatches();
   unsigned int getNumberOfResources();
   bool hasPatches();
-  void registerPatch(PatchDefinition* def);
-  void setDynamicPatchDefinition(PatchDefinition* def){
-    dynamicPatchDefinition = def;
-  }
-  ResourceHeader* getResource(uint8_t index);
-  ResourceHeader* getResource(const char* name);
-  unsigned int getSlot(ResourceHeader* resource);
-  void* getData(ResourceHeader* resource);
-  void store(uint8_t index, uint8_t* data, size_t size);
-  void setDeleted(uint8_t index);
+  /**
+   * @return true if registry has a patch with given 1-based index.
+   */
+  bool hasPatch(uint8_t index);
+  Resource* getPatch(uint8_t index);
+  Resource* getResource(uint8_t index);
+  bool loadProgram(uint8_t slot); // load program from storage
+  bool loadProgram(void* address, uint32_t length); // load program from RAM
 private:
-  bool isPresetBlock(StorageBlock block);
-  StorageBlock patchblocks[MAX_NUMBER_OF_PATCHES];
-  StorageBlock resourceblocks[MAX_NUMBER_OF_RESOURCES];
-  PatchDefinition* defs[MAX_NUMBER_OF_PATCHES];
+  Resource* patches[MAX_NUMBER_OF_PATCHES];
+  Resource* resources[MAX_NUMBER_OF_RESOURCES];
   uint8_t patchCount, resourceCount;
-  PatchDefinition* dynamicPatchDefinition;
+  PatchDefinition patchDefinition;
 };
 
-// Wrappers to be used as callbacks for service calls
-void delete_resource(uint8_t index);
+// // Wrappers to be used as callbacks for service calls
+// void delete_resource(uint8_t index);
 
 #endif // __PatchRegistry_h__