+++ Source/ProgramManager.h
@@ -13,22 +13,11 @@ extern "C" {
   void onResourceUpdate(void);
   void onAudioReady(void);
   void audioCallback(int32_t* rx, int32_t* tx, uint16_t size);
-  /* void encoderChanged(uint8_t encoder, int32_t value); */
-  /* uint16_t getAnalogValue(uint8_t index); */
-  /* void setAnalogValue(uint8_t ch, uint16_t value); */
-
-  /* int16_t getParameterValue(uint8_t ch); */
-  /* void setParameterValue(uint8_t ch, int16_t value); */
-
-  /* uint8_t getButtonValue(uint8_t ch); */
-  /* void setButtonValue(uint8_t ch, uint8_t value); */
-  /* uint16_t getSampleCounter(); */
 }
 
 class ProgramManager {
 private:
   uint8_t patchindex = 0;
-  PatchDefinition* patchdef = NULL;
   void notifyManager(uint32_t ulValue);
   void notifyManagerFromISR(uint32_t ulValue);
 public:
@@ -43,11 +32,6 @@ public:
   void exitProgram(bool isr);
   void resetProgram(bool isr); /* exit and restart program */
   void startProgramChange(bool isr);
-  /* void sendMidiData(int type, bool isr); */
-  
-  // void audioReady();
-  // void programReady();
-  // void programStatus(int);
 
   uint32_t getProgramStackUsed();
   uint32_t getProgramStackAllocation();
@@ -55,19 +39,13 @@ public:
   uint32_t getManagerStackAllocation();
   uint32_t getFreeHeapSize();
 
-  /* void eraseProgramFromFlash(uint8_t sector); */
-  /* void saveProgramToFlash(uint8_t sector, void* address, uint32_t length); */
-  PatchDefinition* getPatchDefinitionFromFlash(uint8_t sector);
-
+  void sendResource(Resource* resource);
   void eraseFromFlash(uint8_t sector);
   void saveToFlash(uint8_t sector, void* address, uint32_t length);
 
   uint32_t getCyclesPerBlock();
   uint32_t getHeapMemoryUsed();
   uint8_t getProgramIndex();
-  PatchDefinition* getPatchDefinition(){
-    return patchdef;
-  }
 private:
   void updateProgramIndex(uint8_t index);
 };