+++ Source/errorhandlers.h
@@ -13,7 +13,6 @@
    const char* getErrorMessage();
    void setErrorStatus(int8_t err);
    void setErrorMessage(int8_t err, const char* msg);
-   void setParameterValues(uint16_t* values, int size);
 
 #define NO_ERROR         0x00
 #define HARDFAULT_ERROR  0x10