+++ HISTORY.md
@@ -1,3 +1,12 @@
+22.0.0
+------
+
+* Prevent rewriting resources and patches
+* Provide resource size and CRC with name
+* Support SPI flash storage
+* Rewritten storage code
+* Support for new Witch device
+
 21.2.2
 ------